+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/DataFeed/DataFeed.cs
@@ -53,7 +53,7 @@ namespace Azure.AI.MetricsAdvisor.Models
             CreatedTime = dataFeedDetail.CreatedTime;
             Creator = dataFeedDetail.Creator;
             IsAdministrator = dataFeedDetail.IsAdmin;
-            MetricIds = dataFeedDetail.Metrics.Select(metric => metric.MetricId).ToList();
+            MetricIds = dataFeedDetail.Metrics.ToDictionary(metric => metric.MetricName, metric => metric.MetricId);
             Name = dataFeedDetail.DataFeedName;
             DataSource = DataFeedSource.GetDataFeedSource(dataFeedDetail);
             SourceType = dataFeedDetail.DataSourceType;
@@ -99,7 +99,7 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// The unique identifiers of the metrics defined in this feed's <see cref="DataFeedSchema"/>.
         /// Set by the service.
         /// </summary>
-        public IReadOnlyList<string> MetricIds { get; }
+        public IReadOnlyDictionary<string, string> MetricIds { get; }
 
         /// <summary>
         /// A custom name for this <see cref="DataFeed"/> to be displayed on the web portal.